edit components relation during searching
